requirements.txt
appdirs==1.4.4
asgiref==3.5.0
astroid==2.12.9
black==21.4b2
bootstrap4==0.1.0
click==7.1.2
dill==0.3.5.1
dj-database-url==1.0.0
Django==4.0.4
django-appconf==1.0.5
django-compressor==4.0
django-heroku==0.3.1
djangorestframework==3.13.1
flake8==3.9.2
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.7.0
pyflakes==2.3.1
pylint==2.15.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2022.2.1
rcssmin==1.1.0
regex==2021.4.4
rjsmin==1.2.0
rope==0.19.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
whitenoise==6.2.0
wrapt==1.14.1
